refactor: 대시보드 데이터 파이프라인 정비 및 사용자 식별 로직 추가
- 하드코딩된 사용자 ID 제거
- 클라이언트 전용 프로젝트 등록 버튼 UI 추가
- 매퍼 인터페이스의 파라미터 타입을 Integer로 변경

+++ src/main/java/com/sanaiclub/project/dao/ProjectBookmarkMapper.java
@@ -5,7 +5,7 @@ import org.apache.ibatis.annotations.Param;
 
 @Mapper
 public interface ProjectBookmarkMapper {
-    int isBookmarked(@Param("projectId") long projectId, @Param("userId") long userId);
-    int insertBookmark(@Param("projectId") long projectId, @Param("userId") long userId);
-    int deleteBookmark(@Param("projectId") long projectId, @Param("userId") long userId);
+    int isBookmarked(@Param("projectId") Integer projectId, @Param("userId") Integer userId);
+    int insertBookmark(@Param("projectId") Integer projectId, @Param("userId") Integer userId);
+    int deleteBookmark(@Param("projectId") Integer projectId, @Param("userId") Integer userId);
 }
